refactor: 의존성 주입 컨테이너 구조 개선 - shared/container에서 환경별 컨테이너를 link-management/di로 이동하여 도메인 중심 구조로 개선, 실행 흐름 명확화
