+++ src/shared/container/service-registry.ts
@@ -1,3 +1,15 @@
+/**
+ * 🔧 서비스 레지스트리 - 공통 의존성 주입 로직
+ *
+ * 환경별 컨테이너(cloudflare-container, local-container)에서
+ * 공통으로 사용하는 서비스 등록 로직을 제공합니다.
+ *
+ * 주요 기능:
+ * - 동적 모듈 로딩
+ * - 서비스 팩토리 패턴
+ * - 의존성 해결 헬퍼
+ */
+
 import { container } from "tsyringe";
 
 // 서비스 의존성 타입 정의
